==> scripts/ppp_sarah/scripts/01_human.Rmd <==
---
title: "humann"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
```

```{r}
meta <- read_csv('../data/phenon.csv', col_names = c('sampleid','pheno')) %>% 
  mutate(sampleid = str_replace(sampleid, '_IGO.+$',''),
         sampleid = str_glue('{sampleid}_humann3')) %>% 
  separate(pheno, into = c('day','grp'), sep = '_C_') %>% 
  mutate(grp = str_replace(grp, '_0.+$', ''))
meta
```

```{r}
full <- read_tsv('../data/humann3_pathabundance_cpm_unstratified.tsv')  %>% 
  rename(pathway = names(.)[1])  

day_pheno <- meta %>% 
  split(., list(.$day))

day_pheno %>% 
  imap(function(.x, .y){
    select(.data = .x, grp) %>% 
      t() %>% 
      write.table(str_glue('../data/lefse/{.y}.txt'), sep = '\t', quote = F, row.names = T, col.names = F)
  })


day_pheno_pcts <- day_pheno %>% 
  map(~ pull(.data = ., sampleid) ) %>% 
  imap(~ full %>% select(pathway, matches(.x)) %>% 
         write_tsv(str_glue('../data/lefse/{.y}_pcts.tsv')))
```
```{r}
fns <- list.files('../data/lefse/', pattern = 'lefse_ready_pcts.tsv$')
    
cmds <- tibble(
  fns = fns
) %>% 
  mutate(format_cmd = str_glue('format_input.py {fns}  {fns}.in -c 1 -u 2')) %>% 
  mutate(run_cmd = str_glue('run_lefse.py {fns}.in  {fns}.res')) %>% 
  mutate(plot_cmd = str_glue('plot_res.py {fns}.res {fns}.pdf --format pdf  --feature_font_size 4 --width 10 --dpi 300 --title {fns}')) %>% 
  select(-fns) %>% 
  gather() %>% 
  select(value) %>% 
  write_csv('../data/lefse/lefse_run_cmd.sh', col_names = F)
```
