==> scripts/ppp_sarah/scripts/02_bile.Rmd <==
---
title: "bile acid KO"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggpubr)
```

```{r}
meta <- read_csv('../data/phenon.csv', col_names = c('sampleid','pheno')) %>% 
  mutate(sampleid = str_replace(sampleid, '_IGO.+$',''),
         #sampleid = str_glue('{sampleid}_humann3')
         ) %>% 
  separate(pheno, into = c('day','grp'), sep = '_C_') %>% 
  mutate(grp = str_replace(grp, '_0.+$', ''))
meta
```


```{r}
dat <- read_tsv('../data/humann3_KO_cpm_joined_unstratified.tsv') %>% 
  rename_all(~ gsub("_humann3", "", .)) %>% 
  rename(KOID = names(.)[1])

KOs <- c('K01442','K15868','K15869','K15870','K15871','K15872','K15873','K15874',
         'K07007','K22604','K22605','K22606','K22607')
KO_str <- paste(KOs, collapse = '|')

tb <- dat %>% 
  filter(str_detect(KOID, KO_str)) %>% 
  gather('sampleid', 'CPM', names(.)[2]:names(.)[ncol(.)]) %>% 
  left_join(meta)

tb %>% 
  write_csv('../data/three_bile_KO_table.csv')

d7_ba_ko <- dat %>% 
  filter(str_detect(KOID, KO_str)) %>% 
  gather('sampleid', 'CPM', names(.)[2]:names(.)[ncol(.)]) %>% 
  left_join(meta) %>% 
  filter(str_detect(KOID, KO_str))  %>% 
  filter(day =='D7')
```
```{r}
d7_ba_ko %>% 
  ggboxplot(x = 'grp', y = 'CPM', add = 'dotplot') +
  facet_grid(day ~ KOID) +
  stat_compare_means(comparisons= list(c('BM', 'BMT')),
            label= "p.value",
            method= 'wilcox.test',
						correct=F) 

ggsave('../data/three_KO_bile_box.pdf')

d7_ba_ko %>% 
  write_csv('../data/02_for_Fig2_three_KO_cpm.csv')
```

```{r}
# have all the counts that come from the bai operon added together (sum of counts for all genes in bai) and then tested for significance
tb %>% 
  group_by(sampleid, day, grp) %>% 
  summarise(total = sum(CPM)) %>% 
  ggboxplot(x = 'grp', y = 'total', add = 'jitter', facet.by = 'day') +
  stat_compare_means(comparisons= list(c('BM', 'BMT')),
            label= "p.value",
            method= 'wilcox.test',
						correct=F) 

ggsave('../data/conbimed_bile_box.pdf')
```

```{r}
# the anti-CD3 projects
kos <- list.files('../data/anticd3/', 'KO', full.names = T)

ko <- kos %>% 
  set_names(kos) %>% 
  map(~ read_tsv(.,  col_types = 'cd') %>% 
          select(KO  = `# Gene Family`, cpm = names(.)[2])  ) %>% 
  bind_rows(.id = 'sampleid') %>% 
  mutate(sampleid = str_replace(sampleid, '../data/anticd3//Sample_',''),
         sampleid = str_replace(sampleid, '_IGO_.+$','')) %>% 
  filter(str_detect(KO, KO_str)) 

ko %>% 
  mutate(grp = if_else(str_detect(sampleid, 'D7_1'), 'control', if_else(str_detect(sampleid, 'D7_2'), 'lower', 'higher'))) %>% 
  filter(!str_detect(KO, '\\|') ) %>% 
  ggboxplot(x = 'grp', 'cpm', facet.by = 'KO', add = 'jitter') +
  facet_grid(KO~., scales = 'free') +
  stat_compare_means(comparisons= list(c('control', 'lower'),
                                       c('control', 'higher'),
                                       c('lower', 'higher')),
										label= "p.format",
										method= 'wilcox.test',
										correct=FALSE)

ggsave('../data/anticd3_KO.pdf', width = 5, height = 6) 
ko %>% write_csv('../data/anticd3_KO.csv')
```
