==> scripts/ppp_sarah/scripts/06_DE_day7.Rmd <==
---
title: "DE methods for day 7 mice data"
author: "Angel"
date: '2022-07-15'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggpubr)
library(vegan)
```

```{r}
pheno <- readxl::read_excel('../data/05_meta_with_alpha.xlsx') %>% 
  filter(day == 'D7')
spp <- read_csv('../data/05_all_shotgun_metaphlan_spp.csv') %>% 
  filter(sampleid %in% pheno$sampleid)
# actually the only difference is the BM and BMT group!
```

```{r}
spp2 <- spp %>% 
  filter(str_detect(clade_name, 'Lactobacillus|Enterococcus'),
         !str_detect(clade_name, 'reuteri'))

spp2 %>% 
  write_csv('../data/06_two_spp_perc.csv')
```


```{r}
genera <- spp %>% 
  mutate(genus = str_replace(clade_name, '\\|s__.+$','')) %>% 
  group_by(sampleid, genus) %>% 
  summarise(relab = sum(perc)) %>% 
  ungroup() %>% 
  mutate(ge = str_extract(genus, 'g__.+$')) %>% 
  #mutate(ge = fct_reorder(ge, genus, )) %>%
  left_join(pheno, by = "sampleid") %>% 
  mutate(sid = str_glue('{grp} {sampleid}')) 
  
genera %>% 
  ungroup() %>% 
  distinct(genus) %>% nrow

keepgenera <- genera %>% 
  filter(relab > 0.1) %>% 
  count(genus) %>% 
  filter(n > floor(nrow(pheno) * 0.2)) %>% 
  pull(genus)

# to find colors for these features
library(randomcoloR)
n <- length(keepgenera)
palette <- distinctColorPalette(n)
pie(rep(1, n), col=palette)

bar_palette <- genera %>%
  filter(genus %in% keepgenera) %>%
  distinct(ge) %>%
  arrange(ge) %>%
  mutate(color = palette)

bar_plot_pal <- deframe(bar_palette)

genera %>% 
  filter(genus %in% keepgenera) %>% 
  ggbarplot(x = 'sid', y = 'relab', fill = 'ge', color = 'grp' , xlab = '', ylab = 'Relative abundance') +
  #facet_wrap(Day ~ fiber, scales = 'free') +
  scale_fill_manual(values = bar_plot_pal) +
  theme( axis.text.x = element_text(angle=90, hjust=0, size = 6.5),
         legend.position = 'bottom') 
```

```{r}
# a beta diversity scatter plot
# find the spp > 0.01% in more than 20% samples
keep_spp <- spp %>% 
  filter(perc > 0.01) %>% 
  count(clade_name) %>% 
  filter(n > floor(nrow(pheno) * 0.2)) %>% 
  pull(clade_name)

# calculate the Bray curtis beta diversity 
cts_fil <- spp %>% 
  filter(clade_name %in% keep_spp) %>% 
  select(sampleid, clade_name,perc ) %>% 
  spread(key = 'clade_name', value = 'perc', fill = 0) %>% 
  column_to_rownames('sampleid')

dist_ <- vegdist(cts_fil, method = 'bray')
eigen <- cmdscale(dist_, eig = T)$eig
percent_var <- signif(eigen/sum(eigen), 3)*100
bc <- cmdscale(dist_, k = 2)
beta_df <- bc %>%
  as.data.frame() %>%
  rownames_to_column('sampleid')  %>% 
  full_join(pheno) 


beta_df %>% 
  ggscatter(x = 'V1', y = 'V2', color =  'grp', palette = 'lancet',alpha = 1, label = 'sampleid') +
  labs(title = '') +
  xlab(paste0("PC 1 [",percent_var[1],"%]")) +
  ylab(paste0("PC 2 [",percent_var[2],"%]")) +
  theme(aspect.ratio=1, legend.position = 'right')

```

```{r}
library(Maaslin2)
spp_cts <- spp %>%   
  filter(clade_name %in% keep_spp) %>% 
  mutate(gspp = str_extract(clade_name, 'g__.+$')) %>% 
  select(sampleid, gspp,perc ) %>% 
  spread(key = 'gspp', value = 'perc', fill = 0) %>% 
  column_to_rownames('sampleid')


fit_data <- Maaslin2(
    input_data = spp_cts, 
    input_metadata = pheno %>% column_to_rownames('sampleid'), 
    normalization = "NONE",
    output = "../data/output", 
    fixed_effects = c("grp"),
    reference = c("grp,BM"))
```

```{r}
# looking at the results from Maaslin2
res <- read_tsv('../data/output/significant_results.tsv')
```
